requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anthropic==0.2.10
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==23.1.0
auth0-python==4.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.6.15
cffi==1.15.1
chainlit==0.2.111
charset-normalizer==2.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.1
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defopt==6.4.0
defusedxml==0.7.1
Deprecated==1.2.14
docutils==0.19
duckdb==0.8.0
et-xmlfile==1.1.0
executing==0.10.0
Faker==18.10.1
fastapi==0.96.0
fastjsonschema==2.17.1
filelock==3.12.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
fvalues==0.0.4
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.31
google-search-results==2.4.2
googleapis-common-protos==1.59.1
gorilla==0.4.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
hnswlib==0.7.0
html5lib==1.1
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.3
importlib-metadata==6.0.1
ipykernel==6.23.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
Kivy==2.1.0
Kivy-Garden==0.1.5
kivy-garden.graph==0.4.1.dev0
kivymd==0.104.2
kiwisolver==1.4.4
langchain==0.0.200
langchain-visualizer==0.0.25
langchainplus-sdk==0.0.10
Levenshtein==0.21.1
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
merge-args==0.1.5
mistune==2.0.5
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.30
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.4
notebook_shim==0.2.3
numerizer==0.2.3
numexpr==2.8.4
numpy==1.23.1
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.0.10
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
ought-ice==0.5.0
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.5.1
plyer==2.0.0
pockets==0.9.1
posthog==3.0.1
prometheus-client==0.17.0
prompt-toolkit==3.0.30
protobuf==4.23.2
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyobjus==1.2.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-Levenshtein==0.21.1
python-socketio==5.8.0
python-ulid==1.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
questionary==1.10.0
rapidfuzz==3.1.1
regex==2022.7.25
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.1
rouge-metric==1.0.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.1.1
scipy==1.9.0
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sphinxcontrib-napoleon==0.7
SQLAlchemy==2.0.15
stack-data==0.4.0
starlette==0.27.0
streamlit==1.23.1
streamlit-chat==0.0.2.2
structlog==23.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
thefuzz==0.19.0
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uptrace==1.18.0
uri-template==1.2.0
urllib3==1.26.11
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchdog-gevent==0.1.1
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wikipedia==1.4.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0